+++ Sources/Sentry/Processors/SentryWatchdogTerminationBreadcrumbProcessor.m
@@ -18,7 +18,13 @@
 @implementation SentryWatchdogTerminationBreadcrumbProcessor
 
 - (instancetype)initWithMaxBreadcrumbs:(NSInteger)maxBreadcrumbs
-                           fileManager:(SentryFileManager *)fileManager
+{
+    return [self initWithMaxBreadcrumbs:maxBreadcrumbs
+                            fileManager:SentryDependencyContainer.sharedInstance.fileManager];
+}
+
+- (instancetype)initWithMaxBreadcrumbs:(NSInteger)maxBreadcrumbs
+                           fileManager:(SentryFileManager *_Nullable)fileManager
 {
     if (self = [super init]) {
         self.fileManager = fileManager;