+++ Sources/Sentry/SentryANRTrackerV2.m
@@ -2,8 +2,8 @@
 
 #if SENTRY_HAS_UIKIT
 
-#    import "SentryDependencyContainer.h"
-#    import "SentryFramesTracker.h"
+#    import "SentryANRStoppedResultInternal.h"
+#    import "SentryANRTrackerInternalDelegate.h"
 #    import "SentryLogC.h"
 #    import "SentrySwift.h"
 #    import "SentryTime.h"
@@ -18,12 +18,12 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
     kSentryANRTrackerStopping
 };
 
-@interface SentryANRTrackerV2 () <SentryANRTracker>
+@interface SentryANRTrackerV2 ()
 
 @property (nonatomic, strong) SentryCrashWrapper *crashWrapper;
 @property (nonatomic, strong) SentryDispatchQueueWrapper *dispatchQueueWrapper;
 @property (nonatomic, strong) SentryThreadWrapper *threadWrapper;
-@property (nonatomic, strong) NSHashTable<id<SentryANRTrackerDelegate>> *listeners;
+@property (nonatomic, strong) NSHashTable<id<SentryANRTrackerInternalDelegate>> *listeners;
 @property (nonatomic, strong) SentryFramesTracker *framesTracker;
 @property (nonatomic, assign) NSTimeInterval timeoutInterval;
 
@@ -32,6 +32,16 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
 @implementation SentryANRTrackerV2 {
     NSObject *threadLock;
     SentryANRTrackerState state;
+}
+
+- (instancetype)initWithTimeoutInterval:(NSTimeInterval)timeoutInterval
+{
+    return
+        [self initWithTimeoutInterval:timeoutInterval
+                         crashWrapper:SentryDependencyContainer.sharedInstance.crashWrapper
+                 dispatchQueueWrapper:SentryDependencyContainer.sharedInstance.dispatchQueueWrapper
+                        threadWrapper:SentryDependencyContainer.sharedInstance.threadWrapper
+                        framesTracker:SentryDependencyContainer.sharedInstance.framesTracker];
 }
 
 - (instancetype)initWithTimeoutInterval:(NSTimeInterval)timeoutInterval
@@ -50,11 +60,6 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
         threadLock = [[NSObject alloc] init];
         state = kSentryANRTrackerNotRunning;
     }
-    return self;
-}
-
-- (id<SentryANRTracker>)asProtocol
-{
     return self;
 }
 
@@ -126,9 +131,9 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
 
             uint64_t framesDelayStartSystemTime = nowSystemTime - appHangStoppedInterval;
 
-            SentryFramesDelayResult *framesDelay =
-                [self.framesTracker getFramesDelay:framesDelayStartSystemTime
-                                endSystemTimestamp:nowSystemTime];
+            SentryFramesDelayResultSPI *framesDelay =
+                [self.framesTracker getFramesDelaySPI:framesDelayStartSystemTime
+                                   endSystemTimestamp:nowSystemTime];
 
             if (framesDelay.delayDuration == -1) {
                 continue;
@@ -172,9 +177,9 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
 
         uint64_t frameDelayStartSystemTime = nowSystemTime - timeoutIntervalInNanos;
 
-        SentryFramesDelayResult *framesDelayForTimeInterval =
-            [self.framesTracker getFramesDelay:frameDelayStartSystemTime
-                            endSystemTimestamp:nowSystemTime];
+        SentryFramesDelayResultSPI *framesDelayForTimeInterval =
+            [self.framesTracker getFramesDelaySPI:frameDelayStartSystemTime
+                               endSystemTimestamp:nowSystemTime];
 
         if (framesDelayForTimeInterval.delayDuration == -1) {
             continue;
@@ -190,7 +195,7 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
 
             reported = YES;
             lastAppHangStartedSystemTime = dateProvider.systemTime;
-            [self ANRDetected:SentryANRTypeFullyBlocking];
+            [self ANRDetected:kSentryANRTypeFullyBlocking];
         }
 
         NSTimeInterval nonFullyBlockingFramesDelayThreshold = self.timeoutInterval * 0.99;
@@ -201,7 +206,7 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
 
             reported = YES;
             lastAppHangStartedSystemTime = dateProvider.systemTime;
-            [self ANRDetected:SentryANRTypeNonFullyBlocking];
+            [self ANRDetected:kSentryANRTypeNonFullyBlocking];
         }
     }
 
@@ -211,15 +216,15 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
     }
 }
 
-- (void)ANRDetected:(enum SentryANRType)type
+- (void)ANRDetected:(SentryANRTypeInternal)type
 {
     NSArray *localListeners;
     @synchronized(self.listeners) {
         localListeners = [self.listeners allObjects];
     }
 
-    for (id<SentryANRTrackerDelegate> target in localListeners) {
-        [target anrDetectedWithType:type];
+    for (id<SentryANRTrackerInternalDelegate> target in localListeners) {
+        [target anrDetected:type];
     }
 }
 
@@ -230,15 +235,15 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
         targets = [self.listeners allObjects];
     }
 
-    SentryANRStoppedResult *result =
-        [[SentryANRStoppedResult alloc] initWithMinDuration:hangDurationMinimum
-                                                maxDuration:hangDurationMaximum];
-    for (id<SentryANRTrackerDelegate> target in targets) {
-        [target anrStoppedWithResult:result];
+    SentryANRStoppedResultInternal *result =
+        [[SentryANRStoppedResultInternal alloc] initWithMinDuration:hangDurationMinimum
+                                                        maxDuration:hangDurationMaximum];
+    for (id<SentryANRTrackerInternalDelegate> target in targets) {
+        [target anrStopped:result];
     }
 }
 
-- (void)addListener:(id<SentryANRTrackerDelegate>)listener
+- (void)addListener:(id<SentryANRTrackerInternalDelegate>)listener
 {
     @synchronized(self.listeners) {
         [self.listeners addObject:listener];
@@ -256,7 +261,7 @@ typedef NS_ENUM(NSInteger, SentryANRTrackerState) {
     }
 }
 
-- (void)removeListener:(id<SentryANRTrackerDelegate>)listener
+- (void)removeListener:(id<SentryANRTrackerInternalDelegate>)listener
 {
     @synchronized(self.listeners) {
         [self.listeners removeObject:listener];