+++ Sources/Sentry/SentryAutoSessionTrackingIntegration.m
@@ -1,9 +1,8 @@
 #import "SentryAutoSessionTrackingIntegration.h"
-#import "SentryDependencyContainer.h"
 #import "SentryLogC.h"
 #import "SentryOptions.h"
 #import "SentrySDKInternal.h"
-#import "SentrySessionTracker.h"
+#import "SentrySwift.h"
 
 NS_ASSUME_NONNULL_BEGIN