+++ Sources/Sentry/SentrySpan.m
@@ -1,7 +1,6 @@
 #import "SentryBaggage.h"
 #import "SentryCrashThread.h"
 #import "SentryDefaultThreadInspector.h"
-#import "SentryDependencyContainer.h"
 #import "SentryFrame.h"
 #import "SentryInternalDefines.h"
 #import "SentryLogC.h"
@@ -17,10 +16,6 @@
 #import "SentryTraceContext.h"
 #import "SentryTraceHeader.h"
 #import "SentryTracer+Private.h"
-
-#if SENTRY_HAS_UIKIT
-#    import <SentryFramesTracker.h>
-#endif // SENTRY_HAS_UIKIT
 
 #if SENTRY_TARGET_PROFILING_SUPPORTED
 #    import "SentryContinuousProfiler.h"
@@ -199,13 +194,6 @@ NS_ASSUME_NONNULL_BEGIN
     }
 }
 
-#if !SDK_V9
-- (void)setExtraValue:(nullable id)value forKey:(NSString *)key
-{
-    [self setDataValue:value forKey:key];
-}
-#endif // !SDK_V9
-
 - (void)removeDataForKey:(NSString *)key
 {
     @synchronized(_data) {
@@ -282,7 +270,7 @@ NS_ASSUME_NONNULL_BEGIN
 #if SENTRY_HAS_UIKIT
     if (_framesTracker.isRunning) {
         CFTimeInterval framesDelay = [_framesTracker
-                getFramesDelay:_startSystemTime
+             getFramesDelaySPI:_startSystemTime
             endSystemTimestamp:SentryDependencyContainer.sharedInstance.dateProvider.systemTime]
                                          .delayDuration;
 
@@ -295,7 +283,9 @@ NS_ASSUME_NONNULL_BEGIN
         NSInteger slowFrames = currentFrames.slow - initSlowFrames;
         NSInteger frozenFrames = currentFrames.frozen - initFrozenFrames;
 
-        if (sentryShouldAddSlowFrozenFramesData(totalFrames, slowFrames, frozenFrames)) {
+        if ([SentryFramesTracker shouldAddSlowFrozenFramesDataWithTotalFrames:totalFrames
+                                                                   slowFrames:slowFrames
+                                                                 frozenFrames:frozenFrames]) {
             [self setDataValue:@(totalFrames) forKey:@"frames.total"];
             [self setDataValue:@(slowFrames) forKey:@"frames.slow"];
             [self setDataValue:@(frozenFrames) forKey:@"frames.frozen"];