+++ Sources/Sentry/include/SentryPrivate.h
@@ -1,6 +1,8 @@
 // Sentry internal headers that are needed for swift code; you cannot import headers that depend on
 // public interfaces here
 #import "NSLocale+Sentry.h"
+#import "SentryANRStoppedResultInternal.h"
+#import "SentryANRTrackerInternalDelegate.h"
 #import "SentryBinaryImageCacheCallbacks.h"
 #import "SentryConcurrentRateLimitsDictionary.h"
 #import "SentryCrashBinaryImageCache.h"
@@ -22,15 +24,23 @@
 
 // Headers that also import SentryDefines should be at the end of this list
 // otherwise it wont compile
+#import "SentryANRTrackerV1.h"
+#import "SentryANRTrackerV2.h"
 #import "SentryAsyncLog.h"
 #import "SentryClient+Logs.h"
+#import "SentryContinuousProfiler.h"
 #import "SentryCrash.h"
 #import "SentryCrashDebug.h"
 #import "SentryCrashMonitor_AppState.h"
 #import "SentryCrashMonitor_System.h"
 #import "SentryDateUtils.h"
+#import "SentryDebugImageProviderInternal.h"
+#import "SentryDefaultAppStateManager.h"
 #import "SentryDefaultThreadInspector.h"
+#import "SentryDefaultUIViewControllerPerformanceTracker.h"
+#import "SentryDelayedFramesTracker.h"
 #import "SentryDependencyContainerSwiftHelper.h"
+#import "SentryDeviceContextKeys.h"
 #import "SentryEvent+Serialize.h"
 #import "SentryFileIOTrackerHelper.h"
 #import "SentryFileManagerHelper.h"
@@ -40,6 +50,7 @@
 #import "SentryMsgPackSerializer.h"
 #import "SentryNSDictionarySanitize.h"
 #import "SentryOptions+Private.h"
+#import "SentryPerformanceTracker.h"
 #import "SentryProfiler+Private.h"
 #import "SentrySDKInternal.h"
 #import "SentryScope+PrivateSwift.h"
@@ -52,4 +63,6 @@
 #import "SentrySysctlObjC.h"
 #import "SentryTraceHeader.h"
 #import "SentryTraceOrigin.h"
+#import "SentryTraceProfiler.h"
 #import "SentryUser+Serialize.h"
+#import "SentryWatchdogTerminationBreadcrumbProcessor.h"