+++ Sources/Swift/Integrations/Performance/SentryUIViewControllerPerformanceTracker.swift
@@ -1,5 +1,6 @@
 #if (os(iOS) || os(tvOS) || (swift(>=5.9) && os(visionOS))) && !SENTRY_NO_UIKIT
 
+@_implementationOnly import _SentryPrivate
 import UIKit
 
 @_spi(Private) @objc public protocol SentryInitialDisplayReporting {
@@ -8,40 +9,80 @@ import UIKit
 
 @_spi(Private) @objc public class SentrySwiftUISpanHelper: NSObject {
     @objc public let hasSpan: Bool
-    @objc public let initialDisplayReporting: SentryInitialDisplayReporting?
     
-    @objc public init(hasSpan: Bool, initialDisplayReporting: SentryInitialDisplayReporting?) {
+    @objc public func reportInitialDisplay() {
+        initialDisplayReporting()
+    }
+    private let initialDisplayReporting: () -> Void
+    
+    @objc public init(hasSpan: Bool, initialDisplayReporting: @escaping () -> Void) {
         self.hasSpan = hasSpan
         self.initialDisplayReporting = initialDisplayReporting
     }
 }
 
-@_spi(Private) @objc public protocol SentryUIViewControllerPerformanceTracker {
+@_spi(Private) @objc public final class SentryUIViewControllerPerformanceTracker: NSObject {
     
-    var inAppLogic: SentryInAppLogic { get set }
+    @objc private let helper: SentryDefaultUIViewControllerPerformanceTracker
     
-    var alwaysWaitForFullDisplay: Bool { get set }
+    override init() {
+        let options = SentrySDKInternal.options
+        inAppLogic = SentryInAppLogic(inAppIncludes: options?.inAppIncludes ?? [], inAppExcludes: options?.inAppExcludes ?? [])
+        helper = SentryDefaultUIViewControllerPerformanceTracker(tracker: SentryPerformanceTracker.shared)
+    }
     
-    func viewControllerLoadView(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void)
+    @objc public var inAppLogic: SentryInAppLogic
     
-    func viewControllerViewDidLoad(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void)
+    @objc public var alwaysWaitForFullDisplay: Bool { get {
+        helper.alwaysWaitForFullDisplay
+    } set {
+        helper.alwaysWaitForFullDisplay = newValue
+    } }
     
-    func viewControllerViewWillAppear(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void)
+    @objc public func viewControllerLoadView(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void) {
+        let inAppLogic = self.inAppLogic
+        helper.viewControllerLoadView(controller, isInApp: { c in
+            inAppLogic.isClassInApp(c)
+        }, callbackToOrigin: callback)
+    }
     
-    func viewControllerViewWillDisappear(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void)
+    @objc public func viewControllerViewDidLoad(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void) {
+        helper.viewControllerViewDidLoad(controller, callbackToOrigin: callback)
+    }
+    
+    @objc public func viewControllerViewWillAppear(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void) {
+        helper.viewControllerViewWillAppear(controller, callbackToOrigin: callback)
+    }
+    
+    @objc public func viewControllerViewWillDisappear(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void) {
+        helper.viewControllerViewWillDisappear(controller, callbackToOrigin: callback)
+    }
 
-    func viewControllerViewDidAppear(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void)
+    @objc public func viewControllerViewDidAppear(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void) {
+        helper.viewControllerViewDidAppear(controller, callbackToOrigin: callback)
+    }
 
-    func viewControllerViewWillLayoutSubViews(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void)
+    @objc public func viewControllerViewWillLayoutSubViews(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void) {
+        helper.viewControllerViewWillLayoutSubViews(controller, callbackToOrigin: callback)
+    }
 
-    func viewControllerViewDidLayoutSubViews(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void)
+    @objc public func viewControllerViewDidLayoutSubViews(_ controller: UIViewController, callbackToOrigin callback: @escaping () -> Void) {
+        helper.viewControllerViewDidLayoutSubViews(controller, callbackToOrigin: callback)
+    }
 
-    func reportFullyDisplayed()
+    @objc public func reportFullyDisplayed() {
+        helper.reportFullyDisplayed()
+    }
     
-    func startTimeToDisplayTracker(
+    @objc public func startTimeToDisplayTracker(
         forScreen screenName: String,
         waitForFullDisplay: Bool,
-        transactionId: SpanId) -> SentrySwiftUISpanHelper
+        transactionId: SpanId) -> SentrySwiftUISpanHelper {
+            let objcType = helper.startTimeToDisplay(forScreen: screenName, waitForFullDisplay: waitForFullDisplay, transactionId: transactionId)
+            return SentrySwiftUISpanHelper(hasSpan: objcType.hasSpan) {
+                objcType.reportInitialDisplay()
+            }
+        }
 }
 
 #endif